requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0

# Advanced Features Dependencies
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.6.0

# AI and ML
openai==1.3.7
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4

# Advanced Search
elasticsearch==8.11.0
django-elasticsearch-dsl==7.3

# Analytics and Monitoring
django-silk==5.0.4
django-debug-toolbar==4.2.0

# Caching and Performance
django-redis==5.4.0
django-compressor==4.4

# Email and Notifications
django-anymail==10.2
sendgrid==6.10.0

# Security and Rate Limiting
django-ratelimit==4.1.0
django-axes==6.1.1
cryptography==41.0.8

# API Documentation
drf-spectacular==0.26.5

# Filter
django-filter==23.3 